requirements.txt
openai==1.12.0
google-generativeai==0.3.2
numpy==1.26.4
tiktoken
transformers
blobfile
torch
sentencepiece
google-cloud-storage